* Moved LauncherLoadingIcon away from the core to the app * cleaned resources, only some of them are still in the core; no duplicate strings anymore
