+++ app/src/main/res/drawable/ol_loading_2.xml
