+++ app/src/main/res/layout/activity_home.xml
@@ -66,7 +66,7 @@
         android:layout_height="match_parent"
         android:background="#90000000">
 
-        <com.benny.openlauncher.core.widget.LauncherLoadingIcon
+        <com.benny.openlauncher.widget.LauncherLoadingIcon
             android:id="@+id/loadingIcon"
             android:layout_width="50dp"
             android:layout_height="50dp"
