+++ core/src/main/res/layout/activity_home.xml
@@ -3,20 +3,6 @@
     android:layout_height="match_parent">
 
     <!-- The main content view -->
-    <include layout="@layout/view_home" />
-
-    <FrameLayout
-        android:id="@+id/loadingSplash"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="#90000000">
-
-        <com.benny.openlauncher.core.widget.LauncherLoadingIcon
-            android:id="@+id/loadingIcon"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:layout_gravity="center" />
-
-    </FrameLayout>
+    <merge layout="@layout/view_home" />
 
 </FrameLayout>
