+++ core/src/main/res/values-ar/strings.xml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <!-- Titles -->
+    <string name="settings">الإعدادات</string>
+
+    <!-- Drag Option View -->
+
+    <string name="edit">تعديل</string>
+    <string name="info">معلومات</string>
+    <string name="uninstall">إلغاء التثبيت</string>
+    <string name="remove">إزالة</string>
+
+    <!-- Desktop Option View -->
+    <string name="home">الرئيسية</string>
+    <string name="action">عمل</string>
+    <string name="lock">قفل</string>
+    <string name="widget">أداة</string>
+
+    <!-- Toasts -->
+
+    <string name="toast_not_enough_space">المساحة غير كافية</string>
+    <string name="toast_app_uninstalled">تم إلغاء تثبيت التطبيق</string>
+
+    <!-- Intro -->
+
+    <!-- Minibar -->
+
+    <!-- About -->
+
+    <!-- Settings -->
+
+    <!-- Preference Keys -->
+
+
+    <!-- Internal Keys -->
+
+
+    <!-- Item Labels -->
+
+    <string name="folder">غير مسمى</string>
+
+    <!-- Dialogs -->
+
+    <string name="cancel">إلغاء الأمر</string>
+    <string name="search_online">البحث عبر الإنترنت</string>
+</resources>
