+++ core/src/main/res/values-ja/strings.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <!-- Titles -->
+    <string name="settings">設定</string>
+
+    <!-- Edit Mode -->
+
+    <string name="info">情報</string>
+    <string name="uninstall">アンインストール</string>
+    <string name="widget">ウィジェット</string>
+    <string name="remove">削除</string>
+
+    <!-- Toasts -->
+
+    <string name="toast_not_enough_space">容量が不足しています</string>
+    <string name="toast_app_uninstalled">このアプリはアンインストールされました。</string>
+
+    <!-- Intro -->
+
+    <!-- MiniBar -->
+
+    <!-- About -->
+
+    <!-- Dialogs -->
+
+    <!-- Settings -->
+
+    <!-- Other -->
+
+    <string name="cancel">キャンセル</string>
+    <string name="home">ホーム</string>
+    <string name="folder">名前なし</string>
+
+</resources>
