+++ core/src/main/res/values-nb-rNO/strings.xml
@@ -1,0 +1,44 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <!-- Titles -->
+    <string name="settings">Innstillinger</string>
+
+    <!-- Drag Option View -->
+
+    <string name="edit">Rediger</string>
+    <string name="info">Info</string>
+    <string name="uninstall">Avinstaller</string>
+    <string name="remove">Fjern</string>
+
+    <!-- Desktop Option View -->
+
+    <string name="home">Hjem</string>
+    <string name="action">Handling</string>
+    <string name="lock">Lås</string>
+    <string name="widget">Miniprogram</string>
+
+    <!-- Toasts -->
+
+    <string name="toast_not_enough_space">Ikke nok plass</string>
+    <string name="toast_app_uninstalled">Program avinstallert.</string>
+
+    <!-- Intro -->
+
+    <!-- Minibar -->
+
+    <!-- About -->
+
+    <!-- Settings -->
+
+    <!-- Preference Keys -->
+
+    <!-- Item Labels -->
+
+    <string name="folder">Uten navn</string>
+
+    <!-- Dialogs -->
+
+    <string name="cancel">Avbryt</string>
+    <string name="search_online">SØK PÅ NETT</string>
+</resources>
