+++ core/src/main/res/values-ru/strings.xml
@@ -1,0 +1,43 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<resources>
+
+    <!-- Titles -->
+
+    <string name="settings">Настройки</string>
+
+   <!-- Drag Option View -->
+
+    <string name="edit">Изменить</string>
+    <string name="info">Информация</string>
+    <string name="uninstall">Удалить</string>
+    <string name="remove">Убрать</string>
+
+   <!-- Desktop Option View -->
+
+   <!-- Toasts -->
+
+    <string name="toast_not_enough_space">Недостаточно места.</string>
+    <string name="toast_app_uninstalled">Приложение удалено.</string>
+
+    <!-- Intro -->
+
+   <!-- Minibar -->
+
+   <!-- About -->
+
+    <!-- Settings -->
+
+   <!-- Preference Keys -->
+
+   <!-- Internal Keys -->
+
+    <!-- Item Labels -->
+
+    <string name="folder">Без имени</string>
+
+    <!-- Dialogs -->
+
+   <string name="cancel">Отмена</string>
+   <string name="search_online">Поиск в сети…</string>
+
+</resources>
